@@ -7,11 +7,17 @@ end
 
 
 # Bundle edge Rails instead: gem 'rails', github: 'rails/rails'
-gem 'rails', '~> 5.1.0'
+gem 'rails', '~> 6.0.0'
+# Reduces boot times through caching; required in config/boot.rb
+gem 'bootsnap', require: false
+gem 'bigdecimal', '~> 1.4.4'
 # Use Puma as the app server
 gem 'puma', '~> 5.2'
 # Use SCSS for stylesheets
-gem 'sass-rails', '~> 5.0'
+gem 'sass', '~> 3.7.4'
+gem 'sass-rails', '~> 5'
+gem 'sprockets', '~> 3.7.2'
+gem 'nokogiri', '~> 1.10.10'
 # Use Uglifier as compressor for JavaScript assets
 gem 'uglifier', '>= 1.3.0'
 # See https://github.com/rails/execjs#readme for more supported runtimes
@@ -21,10 +27,11 @@ gem 'uglifier', '>= 1.3.0'
 gem 'jbuilder', '~> 2.5'
 # Use Redis adapter to run Action Cable in production
 # gem 'redis', '~> 3.0'
-gem 'mysql2'
 
-# Min version 1.8.1 for nokogiri, due to gem vunlerability
-gem 'nokogiri', '>= 1.8.1'
+# Use sqlite3 as the database for Active Record
+gem "sqlite3", "~> 1.4.2"
+
+gem 'mysql2'
 
 group :development, :test do
   # Call 'byebug' anywhere in the code to stop execution and get a debugger console
@@ -35,9 +42,8 @@ group :development, :test do
   gem 'selenium-webdriver', '~> 3.142'
   # For automatically updating chromedriver
   gem 'webdrivers', '~> 4.0', require: false
-  gem 'rspec-rails', '~> 3.5'
-  gem 'factory_girl_rails', ' ~> 4.0'
-  gem 'shoulda-matchers', git: 'https://github.com/thoughtbot/shoulda-matchers.git', branch: 'rails-5'
+  gem 'rspec-rails', '~> 4.0'
+  gem 'factory_bot_rails', ' ~> 4.0'
   gem 'simplecov',      require: false
   gem 'simplecov-lcov', require: false
   gem 'database_cleaner'
@@ -65,9 +71,9 @@ gem 'listen', '>= 3.0.5', '< 3.2'
 
 # Login/auth gems
 gem 'devise'
-gem 'cul_omniauth', '~> 0.6.1'
-gem 'cancancan', '~> 2.0'
+gem 'cul_omniauth', github: "cul/cul_omniauth", branch: "rails_6"
+gem 'cancancan', '~> 3.0'
 
 gem 'gaffe'
 
-gem "cul-ldap", :github => "cul/cul-ldap"
+gem "cul-ldap", github: "cul/cul-ldap"
