@@ -1,9 +1,9 @@
 require 'rails_helper'
 
 describe "Timetables", type: :feature, js: true do
-  let!(:all_location) { FactoryGirl.create(:all_location) }
-  let!(:lehman) { FactoryGirl.create(:lehman) }
-  let!(:miskatonic) { FactoryGirl.create(:miskatonic) }
+  let!(:all_location) { FactoryBot.create(:all_location) }
+  let!(:lehman) { FactoryBot.create(:lehman) }
+  let!(:miskatonic) { FactoryBot.create(:miskatonic) }
 
   shared_examples 'view the page' do
     it "should display a calendar" do
@@ -24,6 +24,7 @@ describe "Timetables", type: :feature, js: true do
       select "06 PM", :from => "timetable_close_4i"
       select "30", :from => "timetable_close_5i"
       click_button("Update Hours")
+      find("td", :text => "16")
       visit(exceptional_edit_location_timetables_path(location_code: lehman.code))
       expect(find("td", :text => "16")).to have_content("7:30AM-6:30PM")
     end
@@ -120,7 +121,7 @@ describe "Timetables", type: :feature, js: true do
   end
 
   describe 'when butler editor logged in' do
-    let(:butler) { FactoryGirl.create(:butler) }
+    let(:butler) { FactoryBot.create(:butler) }
 
     include_context 'login user'
     before :each do
