@@ -36,13 +36,6 @@ Dir[Rails.root.join('spec/support/**/*.rb')].each { |f| require f }
 # If you are not using ActiveRecord, you can remove this line.
 ActiveRecord::Migration.maintain_test_schema!
 
-Shoulda::Matchers.configure do |config|
-  config.integrate do |with|
-    with.test_framework :rspec
-    with.library :rails
-  end
-end
-
 RSpec.configure do |config|
   # Remove this line if you're not using ActiveRecord or ActiveRecord fixtures
   config.fixture_path = "#{::Rails.root}/spec/fixtures"
@@ -56,7 +49,7 @@ RSpec.configure do |config|
   config.include Devise::Test::ControllerHelpers, type: :view
 
   # FactoryGirl helpers.
-  config.include FactoryGirl::Syntax::Methods
+  config.include FactoryBot::Syntax::Methods
 
   # RSpec Rails can automatically mix in different behaviours to your tests
   # based on their file location, for example enabling you to call `get` and
@@ -82,7 +75,7 @@ end
 
 class FindOrCreateStrategy
   def initialize
-    @strategy = FactoryGirl.strategy_by_name(:create).new
+    @strategy = FactoryBot.strategy_by_name(:create).new
   end
 
   delegate :association, to: :@strategy
@@ -101,4 +94,4 @@ class FindOrCreateStrategy
   end
 end
 
-FactoryGirl.register_strategy(:find_or_create, FindOrCreateStrategy)
+FactoryBot.register_strategy(:find_or_create, FindOrCreateStrategy)
