@@ -14,8 +14,9 @@ module LdpdHours
     config.autoload_paths += %W(#{config.root}/lib)
 
     # Settings in config/environments/* take precedence over those specified here.
-    # Application configuration should go into files in config/initializers
-    # -- all .rb files in that directory are automatically loaded.
+    # Application configuration can go into files in config/initializers
+    # -- all .rb files in that directory are automatically loaded after loading
+    # the framework and any gems in your application.
     config.action_dispatch.rescue_responses.merge! 'CanCan::AccessDenied' => :forbidden
 
     config.time_zone = 'Eastern Time (US & Canada)'
